Address build warnings

+++ modules/tracktion_engine/midi/bc_Midi.h
@@ -18,7 +18,7 @@ namespace tracktion {
 
 			struct midiNoteValue
 			{
-				const int midiNote;
+				const unsigned int midiNote;
 				const double freqHz;
 				const juce::String noteName;
 			};
@@ -182,11 +182,11 @@ namespace tracktion {
 							}
 							else if (p_hz < midiNoteValues.begin()->freqHz)
 							{
-								return midiNoteValues.begin()->freqHz;
+								return (int)midiNoteValues.begin()->freqHz;
 							}
 							else if (p_hz > midiNoteValues.back().freqHz)
 							{
-								return midiNoteValues.back().freqHz;
+								return (int)midiNoteValues.back().freqHz;
 							}
 						}
 					}
